pyproject.toml
[tool.poetry]
name = "polymarket-ai-trader"
version = "2.0.0"
description = "Polymarket AI Trader V2 (Beta) - Event-Driven Trading System"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
nautilus-trader = {extras = ["polymarket"], version = "1.202.0"}
google-generativeai = "*"
web3 = ">=7.0.0"
eth-account = ">=0.8.0"
python-dotenv = "^1.0.0"
pandas = "^2.0.0"
numpy = "^1.24.0"
aiohttp = "^3.8.0"
msgspec = "^0.18.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
mypy = "^1.0.0"
black = "^23.0.0"
isort = "^5.12.0"
pytest = "^7.0.0"
pytest-asyncio = "^0.21.0"
types-requests = "^2.31.0"
pandas-stubs = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 127

[tool.black]
line-length = 127
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
warn_return_any = false
warn_unused_ignores = true
show_error_codes = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
